refactor: 공용으로 사용하는 컴포넌트 components/@shared 폴더로 정리
