+++ frontend/src/components/@shared/Animation/NoResult.tsx
