+++ frontend/src/components/@shared/Animation/index.ts
