+++ frontend/src/components/@shared/Card/index.styled.ts
