+++ frontend/src/components/@shared/ErrorBoundary/CategoryFallback.tsx
