+++ frontend/src/components/@shared/ErrorBoundary/index.styled.ts
