+++ frontend/src/components/@shared/Modal/index.styled.ts
