+++ frontend/src/components/@shared/NoResult/index.styled.ts
